# [79_1] Properly caching to speed up search-options and search-parameters