--- a/TeXmacs/progs/source/macro-search.scm
+++ b/TeXmacs/progs/source/macro-search.scm
@@ -177,8 +177,7 @@
       (if std
           (begin
             ;;(display* "Std= " std "\n")
-            (for (x std)
-              (ahash-set! v x #t)))
+            (for (x std) (ahash-set! v x #t)))
           (with def (get-init-tree l)
             ;;(display* "  Def= " def "\n")
             (cond ((tree-is? def 'uninit) (noop))