## What
as title

## Why
to speed up `search-options` and `search-parameters`

## How to test your changes?
Bench `object xmenu= call ("menu-expand", eval ("'" * menu));` in `tm_window_rep::get_menu_widget`

We can find that in beamer style, in the inserted beamer `tit`, menu-expand time of the focus toolbar could be reduced from 96ms to 55ms after the speed up of `search-options`, and it is reduced from 55ms to 25ms after the speed up of `search-parameters`.